+++ app/src/main/java/com/yucwang/turtle/Lesson/LessonFragment.kt
@@ -1,4 +1,4 @@
-package com.yucwang.turtle
+package com.yucwang.turtle.Lesson
 
 
 import android.os.Bundle
@@ -6,7 +6,7 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
+import com.yucwang.turtle.R
 import com.yucwang.turtle.Theme.SystemUiUtils
 import com.yucwang.turtle.Theme.ThemeInterface
